Fix warning when size_t and and int are different sizes, such as on Win64.
git-svn-id: svn+ssh://aegisub/var/svn/aegisub/branches/audio_display_rewrite@4109 93549f3f-7f0a-0410-a4b3-e966c9c94f04

+++ aegisub/src/audio_renderer_spectrum.cpp
@@ -109,7 +109,7 @@ struct AudioSpectrumCacheBlockFactory {
 	/// The filling is delegated to the spectrum renderer
 	float *ProduceBlock(size_t i)
 	{
-		float *res = new float[1<<spectrum->derivation_size];
+		float *res = new float[((size_t)1)<<spectrum->derivation_size];
 		spectrum->FillBlock(i, res);
 		return res;
 	}
